Add RSS Link to header
Add the link to the existing RSS feed (which is at https://plan4better.de/blog/index.xml) to the head so RSS reader can discover it.

Added code based on a mix of https://forestry.io/docs/guides/developing-with-hugo/misc/#rss-feeds (which did not work out of the box) and https://pakstech.com/blog/hugo-rss/.

The rss feed does not show the blogpost author and only the summary, not the full content. But the template comes from forestry.io and I could not figure out who to overwrite/customize it.